refactor(asset): remove repository interface and update import paths
- Deleted the repository interface implementation to streamline asset management.
- Updated import paths in GCS to reflect the new repository structure.
- Enhanced modularity and maintainability by removing legacy components.
